+++ src/renderer/editor/common/item-dragger.ts
@@ -18,15 +18,18 @@ import * as math from '../../math';
  * the `Preview` class. However, also the `Toolbar` uses it to make new menu items
  * draggable.
  *
- * During a drag, the given `div` will have the class `.dragging` added. This can be used
- * to style the `div` differently during a drag.
- *
  * The class is an `EventEmitter` and emits the following events:
  *
  * @fires drag-start - When a drag starts, this event is emitted with the dragged node.
  * @fires drag-move - When a drag moves, this event is emitted with the dragged node and
  *   the current offset.
  * @fires drag-end - When a drag ends, this event is emitted with the dragged node.
+ * @fires click - When a click is detected, this event is emitted with the clicked node.
+ *   When the drag threshold is exceeded, no click event is emitted.
+ * @fires mouse-down - When a mouse down event is detected, this event is emitted with the
+ *   clicked node.
+ * @fires mouse-up - When a mouse up event is detected, this event is emitted with the
+ *   clicked node. It will be emitted before either the `click` or `drag-end` event.
  */
 export class ItemDragger extends EventEmitter {
   /**
@@ -74,6 +77,8 @@ export class ItemDragger extends EventEmitter {
           y: rect.top - parentRect.top,
         };
 
+        this.emit('mouse-down', node, div);
+
         const onMouseMove = (e2: MouseEvent) => {
           const dragCurrent = { x: e2.clientX, y: e2.clientY };
           const offset = math.subtract(dragCurrent, dragStart);
@@ -84,7 +89,6 @@ export class ItemDragger extends EventEmitter {
 
             if (this.draggedItem === null) {
               this.draggedItem = { div, node };
-              div.classList.add('dragging');
               this.emit('drag-start', node, div);
             }
 
@@ -100,10 +104,13 @@ export class ItemDragger extends EventEmitter {
         };
 
         const onMouseUp = (e: MouseEvent) => {
+          this.emit('mouse-up', node, div);
+
           if (this.draggedItem) {
             this.emit('drag-end', this.draggedItem.node, this.draggedItem.div, e.target);
-            this.draggedItem.div.classList.remove('dragging');
             this.draggedItem = null;
+          } else {
+            this.emit('click', node, div);
           }
 
           window.removeEventListener('mousemove', onMouseMove);