+++ docs/changelog.md
@@ -24,10 +24,12 @@ This changelog follows the rules of [Keep a Changelog](http://keepachangelog.com
 #### Added
 
 - A subtle pulse animation to the drop indicator in the menu editor's preview.
+- The possibility to lock item positions in the menu preview of the editor. Locked items cannot be reordered but moved to a fixed angle instead.
 
 #### Changed
 
 - The drop indicator in the menu editor's preview will now move to submenus or the back-navigation button when an item is dragged over them.
+- The mouse cursor will now change to a grabbing hand when moving an item in the menu editor's preview.
 
 
 ## [Kando 0.5.0](https://github.com/kando-menu/kando/releases/tag/v0.5.0)